pyproject.toml
[project]
name = "repl-nix-meetmetricsanalyzer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.10.4",
    "google-api-python-client>=2.148.0",
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.1",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "streamlit>=1.39.0",
]
